Enhances LaTeX project structure and adds new content
Improves the organization of the LaTeX project by modularizing configuration and content into separate files.

Adds support for new commands and required packages to streamline document preparation.

Introduces new sections for introduction and conclusion, along with relevant acknowledgements and abstract in both English and Italian.

Updates the bibliography with recent references and enhances the .gitignore to exclude intermediate files.

Refactors the main document to utilize inputs for better maintainability and readability.